Fix for entities for cards like auto-entities being overridden
Cards like auto-entities were having the entities in their child card overridden by the entities from card-templater. This was the result of code which sets the entities collection of entities and glance cards without entities defined from the templated entities. This fix updates the logic to only do this one card deep (e.g. if you are directly templating an entities card)